extract context/provider to own file
